Extract proof runtime packets
